- Improvements to FitDialog. - Renamed Note::import() method to importASCII() in order to fix conflict with Python rezerved word.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1098 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ fitPlugins/fitRational0/fitRational0.cpp
@@ -13,8 +13,8 @@
 #endif
 
 struct data {
-  size_t n;
-  size_t p;
+  int n;
+  int p;
   double * X;
   double * Y;
   double * sigma;//weighting data
@@ -43,7 +43,7 @@ return (params[1] + x*params[2])/(1 + x*params[0]);
 extern "C" MY_EXPORT int function_f (const gsl_vector * x, void *params,
         gsl_vector * f)
 {
-  size_t n = ((struct data *)params)->n;
+  int n = ((struct data *)params)->n;
   double *X = ((struct data *)params)->X;
   double *Y = ((struct data *)params)->Y;
   double *sigma = ((struct data *)params)->sigma;
@@ -52,7 +52,7 @@ extern "C" MY_EXPORT int function_f (const gsl_vector * x, void *params,
   double b = gsl_vector_get (x, 1);
   double c = gsl_vector_get (x, 2);
 
-  size_t i;
+  int i;
   for (i = 0; i < n; i++)
     {
       double Yi = (b + c*X[i])/(1 + a*X[i]);
@@ -63,7 +63,7 @@ extern "C" MY_EXPORT int function_f (const gsl_vector * x, void *params,
 }
 extern "C" MY_EXPORT double function_d (const gsl_vector * x, void *params)
 {
-  size_t n = ((struct data *)params)->n;
+  int n = ((struct data *)params)->n;
   double *X = ((struct data *)params)->X;
   double *Y = ((struct data *)params)->Y;
   double *sigma = ((struct data *)params)->sigma;
@@ -72,7 +72,7 @@ extern "C" MY_EXPORT double function_d (const gsl_vector * x, void *params)
   double b = gsl_vector_get (x, 1);
   double c = gsl_vector_get (x, 2);
 
-  size_t i;
+  int i;
   double val=0;
   for (i = 0; i < n; i++)
     {
@@ -86,7 +86,7 @@ extern "C" MY_EXPORT double function_d (const gsl_vector * x, void *params)
 extern "C" MY_EXPORT int function_df (const gsl_vector * x, void *params,
          gsl_matrix * J)
 {
-  size_t n = ((struct data *)params)->n;
+  int n = ((struct data *)params)->n;
   double *X = ((struct data *)params)->X;
   double *sigma = ((struct data *)params)->sigma;
 
@@ -94,7 +94,7 @@ extern "C" MY_EXPORT int function_df (const gsl_vector * x, void *params,
   double b = gsl_vector_get (x, 1);
   double c = gsl_vector_get (x, 2);
 
-  size_t i;
+  int i;
   for (i = 0; i < n; i++)
     {
       /* Jacobian matrix J(i,j) = dfi / dxj,
